Test actually collects some data
Text isn't handled yet, but otherwise the test is working. It takes the
XML document representing a figure with lines with points with x,y and
turns it into an Object. I am using 'emit' to get collected data out,
which was not my plan but I think is actually cleaner. Emitting things
whenever they are ready should be easier to connect to other streams.