child() can take an array in case of deep nesting
The document I'm interesting in parsing has egregiously deeply nested
tags, and it would be awkward to chain .child() calls all the way down. So
now child can take an array of tags to enter.

To get the related test to work, collectText also gets called in the
context of the XmlCollector.